// src/models/operations/recover-request.ts
import Document from "../document"
import OperationType from "../../enums/operation-type"

/**
 * Data model representing a public key in the DID Document.
 */
export default interface RecoverRequest {
    type: OperationType
    didSuffix: string
    revealValue: string
    delta: {
        updateCommitment: string
        patches: {
            action: string
            document: Document
        }[]
    }
    signedData: string
}